Refactor FixedLengthRecordDatasetOp
